pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-recommendation-system"
version = "0.1.0"
description = "A modern audio recommendation system using content-based filtering with audio feature extraction"
authors = [{name = "AI Projects", email = "ai@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
    "implicit>=0.7.0",
    "lightfm>=1.17",
    "surprise>=1.1.0",
    "xgboost>=1.7.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "streamlit>=1.25.0",
    "tqdm>=4.65.0",
    "pytest>=7.4.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "pre-commit>=3.3.0",
    "pytest>=7.4.0",
]
tracking = [
    "mlflow>=2.5.0",
    "wandb>=0.15.0",
]
viz = [
    "umap-learn>=0.5.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "S", "T20", "SIM"]
ignore = ["E501", "S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
